Initial implementation to allow step sizes smaller than thread period. Currently only exact matches allowed.

+++ core/fmu-import-export/src/main/resources/c-templates/Fmu.h
@@ -29,7 +29,7 @@ struct PeriodicThreadStatus
 {
 	double period;
 	void (*call)(void);
-	int lastExecuted;
+	double lastExecuted;
 };
 
 extern struct FmiBuffer fmiBuffer;
